requirements.txt
selenium==4.15.2
pandas==2.1.3
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
jobspy==0.29.0
scikit-learn==1.3.2
python-dotenv==1.0.0
webdriver-manager==4.0.1
nltk==3.8.1
numpy==1.25.2
openpyxl==3.1.2